Uploaded arcpublish.py 1.0
This script fetches Thames Water discharge-to-environment (DTE) data, specifically Discharge Current Status data, populates a dataframe with it, makes changes to the format of the frame so that ArcGIS online will not reject it, such as changing attribute field names, character limits, converting from British National Grid to decimal longitude/latitude. Then it generates and saves .csv file from that, and updates a specified published feature layer with the new data held within the csv file. This script needs calling, I've been using a .bat file and Windows Task Scheduler and requires being run in an Anaconda3 environment which the imported modules are contained in. I'll be updating the README to make this clear tomorrow morning.